% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoodscanR-package.R
\docType{data}
\name{spe_test}
\alias{spe_test}
\alias{spe}
\title{Example test spatial transcriptomics data}
\format{
A SpatialExperiment object
}
\usage{
data("spe_test")
}
\value{
A SpatialExperiment object
}
\description{
hoodscanR-package has 1 datasets: \itemize{
\item spe_test  Example test spatial transcriptomics data in
SpatialExperiment format. This test data is randomly subsetting
from the publicly available CosMx non-small cell lung cancer data.
Source data: https://nanostring.com/products/cosmx-spatial-molecular-imager/nsclc-ffpe-dataset/.
}
}
\examples{
data(spe_test)
}
\keyword{internal}
