% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoodscanR-package.R
\docType{data}
\name{spe_test}
\alias{spe_test}
\alias{spe}
\title{Example test data}
\format{
A SpatialExperiment object
}
\usage{
data("spe_test")
}
\value{
A SpatialExperiment object
}
\description{
hoodscanR-package has 1 datasets: \itemize{
\item spe_test  An example data
}
}
\examples{
data(spe_test)
}
\keyword{internal}
